🤖 refactor: extract System1 tool wrapper from AIService.streamMessage()
Extract the ~660-line System1 bash output compaction IIFE from
streamMessage() into a standalone module (system1ToolWrapper.ts).

- New system1ToolWrapper.ts: wrapToolsWithSystem1() + helpers
- Move getTaskDepthFromConfig() to taskUtils.ts (shared location)
- Remove 14 dead imports from aiService.ts
- aiService.ts: 2543 → 1841 lines (−702)
